Refactor datasets app to remove duplicate view and pagination code
