+++ changelog/disable-pagination-offset-cutoff.bugfix.md
@@ -1,0 +1,1 @@
+Unset cursor pagination `offset_cutoff` on dataset endpoints to fix issue when duplicate `created_on` dates are encountered.
