--- datahub/dataset/interaction/pagination.py
@@ -1,7 +1,0 @@
-from rest_framework.pagination import CursorPagination
-
-
-class InteractionsDatasetViewCursorPagination(CursorPagination):
-    """Cursor Pagination for InteractionsDatasetView"""
-
-    ordering = ('created_on', 'pk')
